requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytz>=2023.3
